stage

+++ core/upload_img.py
@@ -9,7 +9,7 @@ def uploadImage(filePath):
         # Save it with the specified quality and output path
         img.save(f"{filePath}.jpg", "JPEG", quality=60)
         # TODO 异步
-        time.sleep(1)
+        time.sleep(0.5)
         with open(f"{filePath}.jpg", 'rb') as file:
             url = 'https://data.test.meituan.com/pc/v1/screenshot/upload'
             headers = {
@@ -30,12 +30,13 @@ def uploadImage(filePath):
                 print(f'Failed to upload file. Status code: {response.status_code}')
 
 directory = settings.SETUP_DIR / "output"
-for filename in os.listdir(directory):
-    file_path = os.path.join(directory, filename)
-    try:
-        if os.path.isfile(file_path) or os.path.islink(file_path):
-            os.unlink(file_path)
-        elif os.path.isdir(file_path):
-            os.rmdir(file_path)
-    except Exception as e:
-        print(f"Failed to delete {file_path}. Reason: {e}")
+if os.path.isdir(directory):
+    for filename in os.listdir(directory):
+        file_path = os.path.join(directory, filename)
+        try:
+            if os.path.isfile(file_path) or os.path.islink(file_path):
+                os.unlink(file_path)
+            elif os.path.isdir(file_path):
+                os.rmdir(file_path)
+        except Exception as e:
+            print(f"Failed to delete {file_path}. Reason: {e}")
